+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 1.9.3 (unreleased)
 ------------------
 
+- Allow authentication with github token
 - Add support for Python 3.9.
-
 - Drop support for Python 3.5.